+++ scripts/autoloads/GameStates/HostGameState.cs
@@ -19,14 +19,14 @@ namespace Game.StateMachines
             return null;
         }
 
-        public override GameState Process(float deltaTime)
+        public override GameState Process(double deltaTime)
         {
             // TODO: Implement host game state
             GD.Print("HostGameState Process");
             return null;
         }
 
-        public override GameState PhysicsProcess(float deltaTime)
+        public override GameState PhysicsProcess(double deltaTime)
         {
             // TODO: Implement host game state
             GD.Print("HostGameState PhysicsProcess");
